new
My name is Omary and here is my story. Maliasiri is a slum area in Dar es Salaam, Tanzania and it is where I call my sweet home. Poverty has always been the biggest struggle in this neighborhood of about 500 adult residents. Who can change the situation? Our watoto ('children' in Swahili), our future.


However, due to the lack of capitals and resources, many of our children unfortunately do not have the opportunity of early education. In fact, among about 300 children in the community, only approximately 50 are sent to nurseries nearby at the moment. Without being educated in a nursery, a child will later face difficulties of getting into a primary school (due to the large number of children enrolling in primary schools here every year, the schools can refuse to receive children who did not received early education). This means, he/she may completely lose the chance of changing his/her life through education. That is why I believe having our children got access to nurseries is the first step of making a change in this community. Building a nursery has been my biggest dream and together with 5 other youngsters we have been working hard to make it come true. Our team plan to continue this work as a community-based organization, we named it 'Mama center Tz'

My family has a land in Maliasili and we have built two small houses as our home there. In front of the houses, there is an empty space where I am planning to build a nursery for our children in the neighborhood. My team and I have already designed the classroom and calculated the costs in details. When the nursery is set up, we should be able to support early education for 50 children. Besides, we have started a entrepreneurship class 4 months ago teaching women and children in the neighborhood a variety of entrepreneurship skills, including batik making, handcrafts, language learning and more. We have educated 36 students so far and sold some products in the community and local markets. At the moment, the class is held everyday in a outdoor space in my home. Once we finish building the nursery, we also hope to continue this class using the space of the classroom.